% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_view_df.R
\name{quickViewObj}
\alias{quickViewObj}
\title{Execute View() on highlighted text in RStudio}
\usage{
quickViewObj(obj)
}
\arguments{
\item{obj}{highlighted_text}
}
\description{
Execute View() on highlighted text in RStudio
}
